Update patch set 4
Patch Set 4: Code-Review+1 Verified+1

This now works, but I am not sure if there is a better way to do it - the current system (as far as I have looked at) requires the torch sysfs entry to be chmodded to 666, meaning any application can enable or disable the torch.

Is this correct and the best way to do it?

Patch-set: 4
Label: Code-Review=+1
Label: Verified=+1